Feature 76: Validations for boilers functions (#68)
* First commit

* boilersRoutes error fixed. Comments added in order to make the last two validations needed

* Validations for create and update functions added. Comments deleted.

* Validation for the remove function added. Errors on previous validations fixed.

* Chenges requested resolved.

* Requested changes resolved on boilersController.

* if-else statements added inside try-catch.

+++ src/controllers/boilersController.js
@@ -7,7 +7,7 @@ const getAllBoilers = async (req, res) => {
     return res.status(200).json(boilers);
   } catch (error) {
     return res.status(500).json({
-      msg: 'An error has occurred',
+      msg: 'An error has occurred.',
     });
   }
 };
@@ -18,88 +18,90 @@ const getBoilerById = async (req, res) => {
 
     if (!boiler) {
       return res.status(400).json({
-        msg: 'The boiler has not been found'
+        msg: 'The boiler has not been found.',
       });
     }
     return res.status(200).json(boiler);
   } catch (error) {
     return res.status(500).json({
-      msg: 'An error has occurred',
+      msg: 'An error has occurred.',
     });
   }
 };
 
 const createBoiler = async (req, res) => {
-  if (!req.body.description
-    || !req.body.boilerType
-    || !req.body.hourMaintenanceCost
-    || !req.body.hourEventualCost
-    || !req.body.maintenanceRate) {
-    return res.status(400).json({
-      msg: 'Error: Missing required fields to create a boiler',
-    });
-  }
-
-  const newBoiler = new models.Boilers({
-    description: req.body.description,
-    boilerType: req.body.boilerType,
-    hourMaintenanceCost: req.body.hourMaintenanceCost,
-    hourEventualCost: req.body.hourEventualCost,
-    maintenanceRate: req.body.maintenanceRate,
-  });
-
   try {
+    const boilerType = await models.BoilerTypes.findById(req.body.boilerType);
+
+    if (!boilerType) {
+      return res.status(400).json({
+        msg: 'The boiler type does not exist in the DataBase.',
+      });
+    }
+    const newBoiler = new models.Boilers({
+      description: req.body.description,
+      boilerType: req.body.boilerType,
+      hourMaintenanceCost: req.body.hourMaintenanceCost,
+      hourEventualCost: req.body.hourEventualCost,
+      maintenanceRate: req.body.maintenanceRate,
+    });
     const result = await newBoiler.save();
 
     return res.status(200).json(result);
   } catch (error) {
     return res.status(500).json({
-      msg: 'An error has occurred',
+      msg: 'An error has occurred.',
     });
   }
 };
 
 const updateBoiler = async (req, res) => {
-  if (!req.body.description
-    || !req.body.boilerType
-    || !req.body.hourMaintenanceCost
-    || !req.body.hourEventualCost
-    || !req.body.maintenanceRate) {
-    return res.status(400).json({
-      msg: 'Error: Missing required fields to update a boilerType',
-    });
-  }
   try {
+    const boilerType = await models.BoilerTypes.findById(req.body.boilerType);
+
+    if (!boilerType) {
+      return res.status(400).json({
+        msg: 'The boiler type does not exist in the DataBase.',
+      });
+    }
     const result = await models.Boilers.findByIdAndUpdate(req.params.id, req.body, { new: true, });
 
     if (!result) {
       return res.status(400).json({
-        msg: 'The boiler has not been found'
+        msg: 'The boiler has not been found.',
       });
     }
     return res.status(200).json(result);
   } catch (error) {
     return res.status(500).json({
-      msg: 'An error has occurred',
+      msg: 'An error has occurred.',
     });
   }
 };
 
 const deleteBoiler = async (req, res) => {
   try {
-    const result = await models.Boilers.findByIdAndDelete(req.params.id);
+    const findBoilerInAppointment = await models.Appointments.findOne({ boiler: req.params.id });
+    const findBoilerInBuildings = await models.Building.findOne({ boilers: req.params.id });
 
-    if (!result) {
-      return res.status(400).json({
-        msg: 'The boiler has not been found'
+    if (!findBoilerInBuildings && !findBoilerInAppointment) {
+      const result = await models.Boilers.findByIdAndDelete(req.params.id);
+
+      if (!result) {
+        return res.status(400).json({
+          msg: 'The boiler has not been found.',
+        });
+      }
+      return res.status(200).json({
+        msg: 'The boiler has been deleted.',
       });
     }
-    return res.status(200).json({
-      msg: 'The boiler has been deleted'
+    return res.status(400).json({
+      msg: 'The boiler is used in an appointment or building.',
     });
   } catch (error) {
     return res.status(500).json({
-      msg: 'An error has occurred'
+      msg: 'An error has occurred.',
     });
   }
 };
